+++ include/save.hpp
@@ -9,8 +9,8 @@ struct SaveFile {
     bool is_new;
     int player_position_x;
     int player_position_y;
-    int player_health = 4 * 4;
-    int player_max_health = 4 * 4;
+    uint8_t player_health = 4 * 4;
+    uint8_t player_max_health = 4 * 4;
 };
 
 struct SaveFiles {
